Updated missing testscript event message and moved check to a separate file.
