+++ controllers/loadtest_controller.go
@@ -18,13 +18,11 @@ import (
 
 	lt "github.com/artilleryio/artillery-operator/api/v1alpha1"
 	"github.com/artilleryio/artillery-operator/internal/telemetry"
-	"github.com/go-logr/logr"
 	"github.com/posthog/posthog-go"
 	v1 "k8s.io/api/batch/v1"
 	core "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -111,29 +109,4 @@ func (r *LoadTestReconciler) SetupWithManager(mgr ctrl.Manager) error {
 		Owns(&v1.Job{}).
 		Owns(&core.Pod{}).
 		Complete(r)
-}
-
-func (r *LoadTestReconciler) ensureTestScriptConfig(ctx context.Context,
-	instance *lt.LoadTest,
-	logger logr.Logger) (*ctrl.Result, error) {
-	configMap := instance.Spec.TestScript.Config.ConfigMap
-
-	found := &core.ConfigMap{}
-	err := r.Get(ctx, types.NamespacedName{
-		Name:      configMap,
-		Namespace: instance.Namespace,
-	}, found)
-
-	if err != nil && errors.IsNotFound(err) {
-		logger.Error(err, "TestScript ConfigMap is missing", "Testscript.Config.ConfigMap", configMap)
-		r.Recorder.Eventf(instance, "Warning", "MissingTestScript", "Cannot locate the Test Script ConfigMap: %s", configMap)
-
-		return &ctrl.Result{}, err
-	} else if err != nil {
-		logger.Error(err, "Failed to get ConfigMap", "Testscript.Config.ConfigMap", configMap)
-		return &ctrl.Result{}, err
-	}
-
-	// ConfigMap located
-	return nil, nil
 }
